+++ packages/expo-file-system/src/next/FileSystem.ts
@@ -19,4 +19,11 @@ export class Directory extends ExpoFileSystem.FileSystemDirectory {
     super(uri);
     this.validatePath();
   }
+
+  list() {
+    // We need to wrap it in the JS File/Directory classes, and returning SharedObjects in lists is not supported yet on Android.
+    return super
+      .listAsRecords()
+      .map(({ isDirectory, path }) => (isDirectory ? new Directory(path) : new File(path)));
+  }
 }