+++ packages/expo-file-system/src/next/FileSystem.types.ts
@@ -52,6 +52,16 @@ export declare class Directory {
    * Moves a directory.
    */
   move(destination: Directory | File);
+  /**
+   * @hidden
+   * Lists the contents of a directory. Should not be used directly, as it returns a list of paths.
+   * This function is internal and will be removed in the future (when returning arrays of shared objects is supported).
+   */
+  listAsRecords(): { isDirectory: string; path: string }[];
+  /**
+   * Lists the contents of a directory.
+   */
+  list(): (Directory | File)[];
 }
 
 /**